requirements.txt
social-auth-app-django==1.1.0
psycopg2==2.6.2
mysqlclient==1.3.9
ipython==5.2.2
djangorestframework==3.5.4
django-debug==0.1.1
django-debug-toolbar==1.6
kombu==3.0.37
redis==2.10.5
celery==3.1.25
